[Feature] Implement Email OTP Authentication (#15)
* Refactor login architecture and modularize UI components

* refactor(core): Centralize navigation and UI components
Moves reusable UI components (`DefaultButton`, `DefaultTextField`, `DefaultDropdowns`) and theme resources from `app` to `core` module.
Introduces `AppNavigator` to manage back stack logic and moves `Route` definitions to `core`.

* feat(auth): Implement modular LoginHandler pattern
Abstracts authentication logic into a generic `LoginHandler` interface with `LoginUiState` and `LoginUiAction`.
Refactors `LoginViewModel` and `LoginScreen` to support dynamic switching between multiple login methods.

* feat(password_login): Extract password authentication module
Creates a new `:password_login` module.
Moves existing username/password logic into `PasswordHandler` and `PasswordLoginLayout`.

* feat(otp_login): Add OTP authentication module
Creates a new `:otp_login` module.
Implements `OtpHandler` and `OtpLayout` for One-Time Password flows.
Adds `requestOtp` and `verifyOtp` methods to `AuthService` and `AuthRepository`.

* chore(di): Update Koin configuration
Registers `AppNavigator`, `PasswordHandler`, and `OtpHandler` in Koin modules.
Injects the list of available login handlers into `LoginViewModel`.

* Refactor navigation logic and update dependencies

* refactor(navigation): Move navigation logic to ViewModels

Migrates navigation handling from UI-side events to `AppNavigator` calls within ViewModels.
- Injects `AppNavigator` into `DashboardViewModel` and `AssetsViewModel`.
- Updates `DashboardViewModel` to navigate to Assets via the navigator instead of emitting `MainAction`.
- Updates `AssetsViewModel` to handle back navigation.
- Updates `AssetsScreen` to invoke the ViewModel for back actions.

* fix(deps): Fix Material Icons Extended dependency

Corrects the version catalog entry and adds the `icons-extended` dependency to the core module.

* style: Code formatting and constants fix

- Removes extra space in `BEARER_TOKEN_PREFIX`.
- Applies standard formatting (spacing, indentation) to `AssetsScreen`, `AssetsViewModel`, and `DashboardViewModel`.
- Updates `MainActivity` indentation.

* Refactor AssetDetail navigation handling

* refactor(asset-detail): Move navigation logic to ViewModel

Updates the `AssetDetail` feature to handle navigation events within the ViewModel using `AppNavigator`.

This includes:
- Creating `AssetDetailAction` sealed interface for UI events.
- Injecting `AppNavigator` into `AssetDetailViewModel`.
- Handling `OnBackClicked` action to navigate back.
- Updating `AssetDetailScreen` to dispatch actions to the ViewModel instead of using a direct callback.

* style: Apply formatting fixes

Reformats code in `MainActivity` and `AssetDetailScreen` for better readability and updates imports.

* Refactor(assets): Update navigation logic and data models

* refactor(assets): Rename and expand AssetsAction
Renames `AssetAction` to `AssetsAction`.
Adds `OnAssetClicked` and `OnBackClicked` events to the interface.
Updates `AssetsViewModel` to handle these events using `AppNavigator`.
Refactors `AssetsScreen` to dispatch actions to the ViewModel instead of using direct navigation callbacks.

* feat(core): Update FloorMap dimension precision
Changes `widthInMeters` and `heightInMeters` from `Int` to `Double` in `FloorMap` and `FloorMapDto` to support precise measurements.
Updates `DashboardScreen` mock data to reflect type changes.

* config: Update network configuration
Updates `BASE_URL` port from 5000 to 5001 in `Constants`.

* feat(auth): Implement request OTP functionality

* Added `SendOtpRequestBody` data model.
* Added `sendOtp` endpoint definition to `AuthApi`.
* Implemented backend integration in `WSAuthService.requestOtp` to replace the stub implementation.

* Implement OTP verification network logic

* Added `VerifyOtpRequestBody` data class for the request payload.
* Added `verifyOtp` endpoint definition to `AuthApi`.
* Updated `WSAuthService` to execute the actual backend call for OTP verification instead of returning a mock result.

* feat(otp): Improve OTP error handling and validation

- **OTP Verification**: Introduced `VerifyOtpThrowable` and `InvalidOtp` exception for specific error handling.
- **Error Handling**: Updated `WSAuthService` to throw `VerifyOtpThrowable.InvalidOtp` on verification failure instead of a generic exception.
- **UI Feedback**: Mapped `InvalidOtp` error to a specific localized string (`invalid_otp_code`) in `OtpHandler`.
- **Validation**: Added email validation in `OtpHandler` before requesting an OTP code.

* refactor(otp): explicit error parameter naming in OtpHandler

Renames the implicit `it` parameter to `error` in the `onFailure` callback within `OtpHandler` for improved code readability.

* feat(otp): Allow changing email address during OTP flow

* Refactor `OtpLayout` to use a reusable `LabelWithTextButton` composable for text-based actions.
* Add a "Wrong e-mail? Change" option to the UI, visible when an OTP has already been sent.
* Add `OnChangeEmailClicked` action to `OtpLoginUiAction`.
* Update `OtpHandler` to reset the login state (email, code, `isOtpSend`) when the change email action is triggered.
* Add string resources for the new UI elements.

* Refactor LabelWithTextButton to core ui component

* refactor(core): Rename DefaultButton.kt and add LabelWithTextButton
Renames `DefaultButton.kt` to `Buttons.kt` and promotes `LabelWithTextButton` to a shared component.

* refactor(password_login): Integrate shared LabelWithTextButton
Replaces local UI logic with the shared `LabelWithTextButton` component and cleans up resource imports.

* refactor(otp_login): Remove redundant LabelWithTextButton
Deletes the local definition of `LabelWithTextButton` in favor of the shared core component.

* feat(login): Add navigation to registration screen

* feat(login): Implement navigation to registration

Adds the ability to navigate to the registration screen from the login screen.
- Updates `LoginViewModel` to inject `AppNavigator` and handle the `OnRegisterClicked` action by navigating to `Route.Registration`.
- Adds `OnRegisterClicked` to the `LoginAction` sealed interface.
- Adds a `LabelWithTextButton` to `LoginScreen` to trigger the registration navigation.

* refactor(otp): Update text and actions in OtpLayout

Updates the UI labels and event handling in the OTP layout.
- Changes the footer label from "Don't have an account" to "Switch login method".
- Renames the action from `OnRegisterClicked` to `OnSwitchLoginMethodClicked`.
- Updates the button label for wrong email scenarios to use the generic "Switch" resource.

* Refactor: remove extra blank line in PasswordHandler

Removes an unnecessary empty line within the `authRepository.login` method call in `PasswordHandler.kt` to clean up formatting.

* feat(navigation): Clear back stack on authentication transitions

Updates navigation logic to remove the previous route from the back stack when transitioning between authentication screens or proceeding to the dashboard.

Specific changes:
- Remove `Route.Login` when navigating to `Route.Registration` or `Route.Dashboard` in `LoginViewModel`, `PasswordHandler`, and `OtpHandler`.
- Remove `Route.Registration` when navigating to `Route.Login` in `RegistrationViewModel`.

* Optimize imports in KoinModules.kt

+++ core/src/main/java/co/be4you/core/data/network/ws/api/models/floorMap/FloorMapDto.kt
@@ -14,7 +14,7 @@ data class FloorMapDto(
     @SerializedName("imageHeightPx")
     val imageHeightPx: Int?,
     @SerializedName("widthInMeters")
-    val widthInMeters: Int,
+    val widthInMeters: Double,
     @SerializedName("heightInMeters")
-    val heightInMeters: Int,
+    val heightInMeters: Double,
 )
